pyproject.toml
[project]
name = "email-service"
version = "0.1.0"
description = "Email service with Celery + Redis using UV"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "pydantic[email]>=2.5.0",
]